requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
google-auth==2.24.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
openai==1.3.5
pydantic==2.5.0
python-multipart==0.0.6
celery==5.3.4
redis==5.0.1
twilio==8.10.0
sendgrid==6.10.0
python-dotenv==1.0.0
apscheduler==3.10.4